did more experimentation. Dicsovered theory issue
Basically, repeated usage of and->or->and->or results in divergence of
the output either to 0 or 1. According to calc_divergence this can
happen in as few as 3 layers in total.

This means some form of normalization is necessary.

2 forms come to mind. The first is a layer that randomly flips several
of the offending bits to force a 50/50 distribution. The other is to
choose activation function (and/or) based off of the distribution of
inputs. Both do essentially the same thing, and most importantly,
require popcounts. :(