worked around zero gradient issue
It seems that the pytorch view() function screws with the order in which
backpropagation happens, allowing custom functions to execute backprop
before all of their dependents are finished. I worked around this by
outputting to a single variable with an extra dimension of size 2, then
unbinding that dimension into 2 variables with the canon torch.unbind()

This is not a solution, just a workaround.

Next step is to look into disappearing gradients